requirements.txt
boto3==1.24.92
garminconnect==0.1.45
python-dotenv==0.21.0